moved architecturs to a generic generators submodule. Also removed test_samplers
