+++ dwave_networkx/__init__.py
@@ -3,21 +3,14 @@ from __future__ import absolute_import
 import sys
 _PY2 = sys.version_info[0] == 2
 
-import dwave_networkx.architectures
-from dwave_networkx.architectures import *
-
-try:
-    import dwave_networkx.x_architectures
-    from dwave_networkx.x_architectures import *
-except ImportError:
-    pass
+import dwave_networkx.generators
+from dwave_networkx.generators import *
 
 import dwave_networkx.algorithms
 from dwave_networkx.algorithms import *
 
 import dwave_networkx.utils
 from dwave_networkx.exceptions import *
-
 
 import dwave_networkx.default_sampler
 from dwave_networkx.default_sampler import *
