+++ dwave_networkx/generators/__init__.py
@@ -1,0 +1,1 @@
+from dwave_networkx.generators.chimera import *
