+++ dwave_networkx/generators/chimera.py
