+++ dwave_networkx/generators/tests/test_chimera.py
