+++ dwave_networkx/utils/__init__.py
@@ -1,3 +1,1 @@
 from dwave_networkx.utils.decorators import *
-
-import dwave_networkx.utils.test_samplers
