+++ requirements.txt
@@ -1,2 +1,3 @@
 networkx>=2.0
 decorator>=4.1.0
+dimod>=0.3.1
